+++ rr/cname.js
@@ -18,8 +18,8 @@ class CNAME extends RR {
     if (net.isIPv4(val) || net.isIPv6(val))
       throw new Error(`CNAME: cname must be a FQDN: RFC 2181`)
 
-    if (!this.fullyQualified('CNAME', 'cname', val)) return
-    if (!this.validHostname('CNAME', 'cname', val)) return
+    if (!this.isFullyQualified('CNAME', 'cname', val)) return
+    if (!this.isValidHostname('CNAME', 'cname', val)) return
     this.set('cname', val)
   }
 
@@ -46,7 +46,7 @@ class CNAME extends RR {
 
     return new this.constructor({
       type     : 'CNAME',
-      name     : fqdn,
+      name     : this.fullyQualify(fqdn),
       cname    : p,
       ttl      : parseInt(ttl, 10),
       timestamp: ts,
@@ -69,7 +69,7 @@ class CNAME extends RR {
   /******  EXPORTERS   *******/
 
   toTinydns () {
-    return `C${this.get('name')}:${this.get('cname')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
+    return `C${this.getTinyFQDN('name')}:${this.get('cname')}:${this.getTinydnsPostamble()}\n`
   }
 }