+++ rr/ns.js
@@ -11,8 +11,8 @@ class NS extends RR {
   setDname (val) {
     if (!val) throw new Error(`NS: dname is required: ${this.getRFCs()}`)
 
-    this.fullyQualified('NS', 'dname', val)
-    this.validHostname('NS', 'dname', val)
+    this.isFullyQualified('NS', 'dname', val)
+    this.isValidHostname('NS', 'dname', val)
 
     this.set('dname', val)
   }
@@ -41,8 +41,8 @@ class NS extends RR {
 
     return new this.constructor({
       type     : 'NS',
-      name     : fqdn,
-      dname    : dname,
+      name     : this.fullyQualify(fqdn),
+      dname    : this.fullyQualify(dname),
       ttl      : parseInt(ttl, 10),
       timestamp: ts,
       location : loc !== '' && loc !== '\n' ? loc : '',
@@ -64,7 +64,7 @@ class NS extends RR {
 
   /******  EXPORTERS   *******/
   toTinydns () {
-    return `&${this.get('name')}::${this.get('dname')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
+    return `&${this.getTinyFQDN('name')}::${this.getTinyFQDN('dname')}:${this.getTinydnsPostamble()}\n`
   }
 }