+++ rr/ptr.js
@@ -8,8 +8,8 @@ class PTR extends RR {
 
   /****** Resource record specific setters   *******/
   setDname (val) {
-    this.fullyQualified('PTR', 'dname', val)
-    this.validHostname('PTR', 'dname', val)
+    this.isFullyQualified('PTR', 'dname', val)
+    this.isValidHostname('PTR', 'dname', val)
 
     this.set('dname', val)
   }
@@ -37,8 +37,8 @@ class PTR extends RR {
 
     return new this.constructor({
       type     : 'PTR',
-      name     : fqdn,
-      dname    : p,
+      name     : this.fullyQualify(fqdn),
+      dname    : this.fullyQualify(p),
       ttl      : parseInt(ttl, 10),
       timestamp: ts,
       location : loc !== '' && loc !== '\n' ? loc : '',
@@ -59,7 +59,7 @@ class PTR extends RR {
 
   /******  EXPORTERS   *******/
   toTinydns () {
-    return `^${this.get('name')}:${this.get('dname')}:${this.getEmpty('ttl')}:${this.getEmpty('timestamp')}:${this.getEmpty('location')}\n`
+    return `^${this.getTinyFQDN('name')}:${this.getTinyFQDN('dname')}:${this.getTinydnsPostamble()}\n`
   }
 }