+++ test/naptr.js
@@ -8,7 +8,7 @@ const NAPTR = require('../rr/naptr')
 const validRecords = [
   {
     class      : 'IN',
-    name       : 'cid.urn.arpa',
+    name       : 'cid.urn.arpa.',
     type       : 'NAPTR',
     order      : 100,
     preference : 10,
@@ -17,7 +17,7 @@ const validRecords = [
     regexp     : '!^urn:cid:.+@([^\\.]+\\.)(.*)$!\2!i',
     replacement: '.',
     ttl        : 86400,
-    testB      : 'cid.urn.arpa\t86400\tIN\tNAPTR\t100\t10\t""\t""\t"!^urn:cid:.+@([^\\.]+\\.)(.*)$!\x02!i"\t.\n',
+    testB      : 'cid.urn.arpa.\t86400\tIN\tNAPTR\t100\t10\t""\t""\t"!^urn:cid:.+@([^\\.]+\\.)(.*)$!\x02!i"\t.\n',
     testT      : ':cid.urn.arpa:35:\\000\\144\\000\\012\\000\\000\\040!^urn\\072cid\\072.+@([^\\134.]+\\134.)(.*)$!\x02!i\\001.``000:86400::\n',
   },
 ]