+++ test/smimea.js
@@ -8,14 +8,14 @@ const SMIMEA = require('../rr/tlsa')
 const validRecords = [
   {
     class                         : 'IN',
-    name                          : '_443._tcp.www.example.com',
+    name                          : '_443._tcp.www.example.com.',
     type                          : 'SMIMEA',
     ttl                           : 3600,
     'certificate usage'           : 0,
     selector                      : 0,
     'matching type'               : 1,
     'certificate association data': '( d2abde240d7cd3ee6b4b28c54df034b9 7983a1d16e8a410e4561cb106618e971 )',
-    testB                         : '_443._tcp.www.example.com\t3600\tIN\tSMIMEA\t0\t0\t1\t( d2abde240d7cd3ee6b4b28c54df034b9 7983a1d16e8a410e4561cb106618e971 )\n',
+    testB                         : '_443._tcp.www.example.com.\t3600\tIN\tSMIMEA\t0\t0\t1\t( d2abde240d7cd3ee6b4b28c54df034b9 7983a1d16e8a410e4561cb106618e971 )\n',
     // testT                      : '',
   },
 ]