+++ test/srv.js
@@ -8,14 +8,14 @@ const SRV = require('../rr/srv')
 const validRecords = [
   {
     class   : 'IN',
-    name    : '_imaps._tcp.example.com',
+    name    : '_imaps._tcp.example.com.',
     type    : 'SRV',
     target  : 'mail.example.com.',
     priority: 1,
     weight  : 0,
     port    : 993,
     ttl     : 3600,
-    testB   : '_imaps._tcp.example.com\t3600\tIN\tSRV\t1\t0\t993\tmail.example.com.\n',
+    testB   : '_imaps._tcp.example.com.\t3600\tIN\tSRV\t1\t0\t993\tmail.example.com.\n',
     testT   : ':_imaps._tcp.example.com:33:\\000\\001\\000\\000\\003\\341\\004mail\\007example\\003com\\000:3600::\n',
   },
 ]