+++ CHANGES.md
@@ -1,6 +1,23 @@
 
 #### 1.N.N - YYYY-MM-DD
 
+
+#### 0.9.3 - 2022-03-22
+
+- hasValidLabels: remove trailing dot, else split returns empty string
+- rename fullyQualified -> isFullyQualified
+- rename validHostname -> isValidHostname
+- set('type') no longer falls back on constructor.name (didn't reliably inherit)
+- new fns: getTinyFQDN, fullyQualify, getPrefix, getFQDN
+- when loading RR classes, ignore files that don't end with .js
+- rr/txt: support data as array (improves idempotency)
+- fromTinydns: fully qualify hostnames
+- toTinydns: strip trailing . upon export
+- rename getCommonFields -> getPrefixFields
+- TXT: import BIND format w/o mangling WS
+- SPF inherits from TXT
+
+
 #### 0.9.2 - 2022-03-18
 
 - mx: weight -> preference
@@ -11,7 +28,7 @@
 - on `index.is*` functions which throw, use declaratively
 - class: add NONE and ANY
 - validHostname: allow / char
-- use \__dirname to find RR mods
+- use \_\_dirname to find RR mods
 
 
 #### 0.9.1 - 2022-03-14